Do not addend vendor dbx in MokListX(RT) mirroring
Our vendor dbx is 19KB; which is huge and causing a lot of
laptops to run out of EFI variable storage space, or MacBooks
will just hang, when mirroring to MokListXR

Our kernels currently do not read MokListXRT anyway, so not
mirroring the vendor dbx does not affect security of current
kernels. Newer kernels do support vendor dbx, but we can just
bake in the vendor dbx into the kernel images too, rather than
rely on an EFI variable to pass that info through to the kernel.

The kernel could even be told to read MokList{,X} directly, before
calling ExitBootServices, in which case we could stop mirroring
entirely.